requirements.txt
django
djangorestframework
Pillow
python-dotenv
djangorestframework-simplejwt